Use int64_t tmp before ModP

+++ smoothquant/c/mod.c
@@ -1,10 +1,10 @@
 #include "mod.h"
 
-int64_t ModP(int64_t x) {
+int32_t ModP(int64_t x) {
     int64_t mod = ((x + P) % (2 * P)) - P;
     if (mod < -P) {
         mod += 2 * P;
     }
-    return mod;
+    return (int32_t)mod;
 }
